refactor(shared): move toPoolerUrl to the pure util lib
