+++ libs/shared/util/pure/src/index.ts
@@ -7,3 +7,4 @@ export * from './lib/get-sentry-sample-rate';
 export * from './lib/is-object';
 export * from './lib/is-regex-pattern';
 export * from './lib/sanitize-svg';
+export * from './lib/to-pooler-url';
